+++ packages/sampler/stories/moonstone-stories/ProgressBar.js
@@ -1,7 +1,7 @@
 import ProgressBar, {ProgressBarBase} from '@enact/moonstone/ProgressBar';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {boolean, number} from '@storybook/addon-knobs';
+import {boolean, number, select} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
 import nullify from '../../src/utils/nullify.js';
@@ -19,7 +19,7 @@ storiesOf('Moonstone', module)
 			<ProgressBar
 				backgroundProgress={number('backgroundProgress', 0.5, {range: true, min: 0, max: 1, step: 0.01})}
 				progress={number('progress', 0.4, {range: true, min: 0, max: 1, step: 0.01})}
-				vertical={boolean('vertical', false)}
+				orientation={select('orientation', ['horizontal', 'vertical'], 'horizontal')}
 				disabled={nullify(boolean('disabled', false))}
 			/>
 		))